Refactor components divide
2강 React - 2주 차 과제 1.
- index.jsx에서 컴포넌트 분리하기 작업.
