+++ src/NumberButton.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function NumberButton({ number, onClick }) {
+  return (
+    <button type="button" onClick={() => onClick(number)}>
+      {number}
+    </button>
+  );
+}
+
+export default NumberButton;
